# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /customer/{id}:
    # our controller name
    x-swagger-router-controller: customer
    get:
      operationId: getOne
      description: get a customer
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/getCustomerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: update
      description: update customer
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          description: The id of the customer
          required: true
        - name: title
          in: body
          description: The name of the customer
          required: true
          schema: 
            $ref: "#/definitions/Customer"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/saveCustomerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: delcustomer
      description: delete customer
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          description: The id of the customer
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /customerclear:
    # our controller name
    x-swagger-router-controller: customer
    delete:
      operationId: clearcustomer
      description: clear the customers list
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

# complex objects have schema definitions
definitions:
  Customer:
    required:
      - name
      - dob
      - gender
    properties:
      name:
        type: string
      dob:
        type: string
      gender:
        type: string
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
      description:
        type: string
    required:
      - success
      - description
  saveCustomerResponse:
    type: object
    properties:
      success:
        type: number
      id:
        type: string
      description:
        type: string
    required:
      - success
      - description
  getCustomerResponse:
    required:
      - id
      - name
      - dob
      - gender
    properties:
      id:
        type: string
      name: 
        type: string
      dob:
        type: string
      gender:
        type: string
...